Dynamic Incubator - Bug 200045 - 1.1.0 Branch Refactored DynamicClassLoader out of DynamicConversionManager to support sharing one between sessions (MOXy & JPA/ORM) Added RelationalMappingFactory to simplify creating types on the fly Added JPAMappingFactory which extends RelationalMappingFactory to support CMP3Policy usage Adjusted all testing for new classes

+++ dynamic/branches/1.1.0/eclipselink.jpa.test.dynamic/src/testing/simple/SimpleType.java
@@ -2,19 +2,16 @@ package testing.simple;
 
 import static junit.framework.Assert.*;
 
-import java.util.*;
+import java.util.Calendar;
 
 import javax.persistence.*;
 
 import org.eclipse.persistence.descriptors.ClassDescriptor;
 import org.eclipse.persistence.dynamic.*;
-import org.eclipse.persistence.exceptions.DatabaseException;
 import org.eclipse.persistence.internal.dynamic.EntityTypeImpl;
-import org.eclipse.persistence.internal.helper.DynamicConversionManager;
 import org.eclipse.persistence.jpa.JpaHelper;
 import org.eclipse.persistence.sessions.IdentityMapAccessor;
 import org.eclipse.persistence.sessions.server.Server;
-import org.eclipse.persistence.tools.schemaframework.DynamicSchemaManager;
 import org.junit.*;
 
 public class SimpleType {
@@ -30,6 +27,8 @@ public class SimpleType {
 
         EntityTypeImpl simpleType = (EntityTypeImpl) DynamicHelper.getType(session, "Simple");
         assertNotNull("'Simple' EntityType not found", simpleType);
+        
+        assertEquals(2, simpleType.getMappingsRequiringInitialization().size());
 
         assertEquals(descriptor, simpleType.getDescriptor());
     }
@@ -37,27 +36,21 @@ public class SimpleType {
     @Test
     public void simpleInstance_CRUD() {
         Server session = JpaHelper.getServerSession(emf);
-        EntityTypeImpl simpleType = (EntityTypeImpl) DynamicHelper.getType(session, "Simple");
-
-        EntityManager em = emf.createEntityManager();
+        IdentityMapAccessor cache = session.getIdentityMapAccessor();
 
         DynamicEntity simpleInstance = createSimpleInstance(emf, 1);
+        assertNotNull(simpleInstance);
 
-        int simpleCount = ((Number) em.createQuery("SELECT COUNT(s) FROM Simple s").getSingleResult()).intValue();
-        assertEquals(1, simpleCount);
-
-        IdentityMapAccessor cache = session.getIdentityMapAccessor();
         assertTrue(cache.containsObjectInIdentityMap(simpleInstance));
-
-        em.clear();
         cache.initializeAllIdentityMaps();
+        assertFalse(cache.containsObjectInIdentityMap(simpleInstance));
 
-        DynamicEntity findResult = (DynamicEntity) em.find(simpleType.getJavaClass(), 1);
-
-        assertNotNull(findResult);
-        assertEquals(simpleInstance.get("id"), findResult.get("id"));
-        assertEquals(simpleInstance.get("value2"), findResult.get("value2"));
-
+        EntityManager em = emf.createEntityManager();
+        
+        assertFalse(em.contains(simpleInstance));
+        simpleInstance =  em.merge(simpleInstance);
+        assertTrue(em.contains(simpleInstance));
+        
         em.close();
     }
 
@@ -101,66 +94,64 @@ public class SimpleType {
         simpleInstance.set("id", 1);
         simpleInstance.set("value2", true);
 
+        assertEquals(0, ((Number) em.createQuery("SELECT COUNT(s) FROM Simple s").getSingleResult()).intValue());
+
         em.getTransaction().begin();
         em.persist(simpleInstance);
         em.getTransaction().commit();
 
+        assertEquals(1, ((Number) em.createQuery("SELECT COUNT(s) FROM Simple s").getSingleResult()).intValue());
+
+        DynamicEntity foundEntity = (DynamicEntity) em.find(simpleEntityType.getJavaClass(), 1);
+        
+        assertNotNull(foundEntity);
+        assertEquals(simpleInstance.get("id"), foundEntity.get("id"));
+        assertEquals(simpleInstance.get("value1"), foundEntity.get("value1"));
+        assertEquals(simpleInstance.get("value2"), foundEntity.get("value2"));
+
         em.close();
+        
         return simpleInstance;
     }
 
     @BeforeClass
     public static void setUp() {
-        Map properties = new HashMap();
-
-        properties.put("eclipselink.ddl-generation.output-mode", "database");
-        properties.put("eclipselink.ddl-generation", "drop-and-create-tables");
-
-        emf = Persistence.createEntityManagerFactory("empty", properties);
-
+        emf = Persistence.createEntityManagerFactory("empty");
         Server session = JpaHelper.getServerSession(emf);
-        try {
-            session.executeNonSelectingSQL("DROP TABLE SIMPLE_TYPE CASCADE CONSTRAINTS");
-        } catch (DatabaseException dbe) {
-            // ignore
-        }
 
-        Class simpleTypeClass = DynamicConversionManager.lookup(session).createDynamicClass("model.Simple");
-        EntityTypeImpl entityType = new EntityTypeImpl(simpleTypeClass, "SIMPLE_TYPE");
-        entityType.addDirectMapping("id", int.class, "SID", true);
-        entityType.addDirectMapping("value1", String.class, "VAL_1", false);
-        entityType.addDirectMapping("value2", boolean.class, "VAL_2", false);
-        entityType.addDirectMapping("value3", Calendar.class, "VAL_3", false);
-        entityType.addDirectMapping("value4", Character.class, "VAL_4", false);
+        RelationalMappingFactory factory = new RelationalMappingFactory(session, "model.Simple", "SIMPLE_TYPE");
+        factory.addDirectMapping("id", int.class, "SID", true);
+        factory.addDirectMapping("value1", String.class, "VAL_1", false);
+        factory.addDirectMapping("value2", boolean.class, "VAL_2", false);
+        factory.addDirectMapping("value3", Calendar.class, "VAL_3", false);
+        factory.addDirectMapping("value4", Character.class, "VAL_4", false);
 
-        entityType.addToSession(session);
-        new DynamicSchemaManager(session).createTables(entityType);
+        factory.addToSession(session, true);
     }
 
     @Before
     @After
     public void clearSimpleTypeInstances() {
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        em.createQuery("DELETE FROM Simple").executeUpdate();
-        em.getTransaction().commit();
-        em.close();
+        if (emf != null && emf.isOpen()) {
+            EntityManager em = emf.createEntityManager();
+            em.getTransaction().begin();
+            em.createQuery("DELETE FROM Simple").executeUpdate();
+            em.getTransaction().commit();
+            em.close();
+        }
     }
 
     @AfterClass
     public static void shutdown() {
-        EntityManager em = emf.createEntityManager();
+        if (emf != null && emf.isOpen()) {
+            EntityManager em = emf.createEntityManager();
 
-        em.getTransaction().begin();
-        em.createNativeQuery("DELETE FROM CUSTOM_REL_MTM").executeUpdate();
-        em.createNativeQuery("DELETE FROM CUSTOM_REL_MTO").executeUpdate();
-        em.createNativeQuery("DELETE FROM CUSTOM_REL_OTO").executeUpdate();
-        em.createNativeQuery("DELETE FROM CUSTOM_FIELD").executeUpdate();
-        em.createNativeQuery("DELETE FROM CUSTOM_TYPE").executeUpdate();
-        em.createNativeQuery("DROP TABLE SIMPLE_TYPE CASCADE CONSTRAINTS").executeUpdate();
-        em.getTransaction().commit();
+            em.getTransaction().begin();
+            em.createNativeQuery("DROP TABLE SIMPLE_TYPE CASCADE CONSTRAINTS").executeUpdate();
+            em.getTransaction().commit();
 
-        em.close();
-        emf.close();
+            em.close();
+            emf.close();
+        }
     }
 }
